v.052

+++ Controllers/OrdersController.cs
@@ -6,6 +6,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.CodeAnalysis;
 using Microsoft.EntityFrameworkCore;
 using System;
+using System.Collections.Generic;
 using System.Linq;
 using static NuGet.Packaging.PackagingConstants;
 
@@ -48,11 +49,11 @@ namespace Thrift_E.Controllers
                     _context.OrderdProducts.Add(OrderdProduct);
                     _context.Carts.Remove(Item);
                 }
-                else 
+                else
                 {
                     return RedirectToAction("Index", "Carts");
                 }
-                
+
             }
             _context.SaveChanges();
             return RedirectToAction("Index");
@@ -63,10 +64,7 @@ namespace Thrift_E.Controllers
             var person = _context.Customers.FirstOrDefault(x => x.Cookie == myCookieValue);
 
             List<Order> orders = _context.Orders.Where(x => x.CustomerId == person.CustomerId).ToList();
-
             return View(orders);
-
-
         }
 
         public IActionResult Details(int OrderId)
@@ -74,7 +72,7 @@ namespace Thrift_E.Controllers
             string myCookieValue = HttpContext.Request.Cookies["MyCookie"];
             var person = _context.Customers.FirstOrDefault(x => x.Cookie == myCookieValue);
 
-            var products = _context.OrderdProducts.Where(z => z.OrderId == OrderId).Include(c => c.Product).ThenInclude(l => l.Category).Include(p => p.Product.MeasureOfScale).Include(p => p.Order).Select(c => new OrderViewModel
+            var products = _context.OrderdProducts.Where(z => z.OrderId == OrderId).Include(c => c.Product).ThenInclude(l => l.Category).Include(p => p.Product.MeasureOfScale).Include(x => x.Product.Brand).Include(p => p.Order).Select(c => new OrderViewModel
             {
                 CustomerId = c.Order.CustomerId,
                 CustomerName = person.FirstName,
@@ -87,14 +85,15 @@ namespace Thrift_E.Controllers
                 CategoryName = c.Product.Category.CategoryName,
                 MeasureOfScaleName = c.Product.MeasureOfScale.MeasureOfScale,
                 Description = c.Product.Description,
+                OrderStatus = c.Order.OrderStatus,
+                Brand = c.Product.Brand.BrandName,
+                OrderDate = c.Order.OrderDate,
+                Size = c.Product.Size,
 
             }).ToList();
             return View(products);
         }
-        public IActionResult Confirm()
-        {
-            return View();
-        }
+
         [HttpPost]
         public IActionResult Confirm1(string Area, string Streat, string House)
         {
@@ -114,29 +113,70 @@ namespace Thrift_E.Controllers
             else order.House = House;
             order.OrderStatus = "Being Processed";
             _context.Orders.Add(order);
-
-            foreach (Cart Item in carts)
+            if (carts.Count != 0)
             {
-                Product product = _context.Products.FirstOrDefault(x => x.ProductId == Item.ProductId);
-                if (product.InstockQty >= Item.Qty)
+                foreach (Cart Item in carts)
                 {
-                    OrderdProduct OrderdProduct = new OrderdProduct();
-                    OrderdProduct.OrderId = order.OrderId;
-                    OrderdProduct.ProductId = Item.ProductId;
-                    OrderdProduct.Qty = (int)Item.Qty;
-                    OrderdProduct.Total = (product.Price * OrderdProduct.Qty);
-                    product.InstockQty -= OrderdProduct.Qty;
-                    _context.OrderdProducts.Add(OrderdProduct);
-                    _context.Carts.Remove(Item);
+                    Product product = _context.Products.FirstOrDefault(x => x.ProductId == Item.ProductId);
+                    if (product.InstockQty >= Item.Qty)
+                    {
+                        OrderdProduct OrderdProduct = new OrderdProduct();
+                        OrderdProduct.OrderId = order.OrderId;
+                        OrderdProduct.ProductId = Item.ProductId;
+                        OrderdProduct.Qty = (int)Item.Qty;
+                        OrderdProduct.Total = (product.Price * OrderdProduct.Qty);
+                        product.InstockQty -= OrderdProduct.Qty;
+                        _context.OrderdProducts.Add(OrderdProduct);
+                        _context.Carts.Remove(Item);
+                    }
+                    else
+                    {
+                        return RedirectToAction("Index", "Carts");
+                    }
                 }
-                else
-                {
-                    return RedirectToAction("Index", "Carts");
-                }
-
+                _context.SaveChanges();
+                return RedirectToAction("Index");
             }
-            _context.SaveChanges();
-            return RedirectToAction("Index");
+            else 
+            {
+                return RedirectToAction("Index", "Carts");
+            }
         }
+        [Authorize(policy: "MustBeAdmin")]
+        public IActionResult ChangeStatus(string Status, int OrderId)
+        {
+            var order = _context.Orders.FirstOrDefault(x => x.OrderId == OrderId);
+            order.OrderStatus = Status;
+            _context.Orders.Update(order);
+            _context.SaveChanges();
+            return RedirectToAction("AllOrders");
+        }
+        [Authorize(policy: "MustBeAdmin")]
+        public IActionResult Delete(int OrderId)
+        {
+            var order = _context.Orders.FirstOrDefault(x => x.OrderId == OrderId);
+            List<OrderdProduct> orderdproducts = _context.OrderdProducts.Where(x => x.OrderId == OrderId).ToList();
+            _context.OrderdProducts.RemoveRange(orderdproducts);
+            _context.Orders.Remove(order);
+            _context.SaveChanges();
+            return RedirectToAction("AllOrders");
+        }
+        [Authorize(policy: "MustBeAdmin")]
+        public IActionResult AllOrders()
+        {
+            var orders = _context.Orders.Include(x => x.Customer).Select(c => new OrderViewModel
+            {
+                CustomerName = c.Customer.FirstName + " " + c.Customer.MiddleName + " " + c.Customer.LastName,
+                Phone = c.Customer.Phone,
+                OrderId = c.OrderId,
+                OrderStatus = c.OrderStatus,
+                OrderDate = c.OrderDate,
+                Location = c.Area + " / " + c.Streat + " / " + c.House,
+
+            }).ToList();
+
+            return View(orders);
+        }
+
     }
 }
